Make load_dotenv usage optional and configurable
- Add environment.use_dotenv config option for Docker/Linux compatibility
- Move model classes from deepeval_integration to testing_model.models
- Update auth_utils to use conditional environment loading
- Restructure imports and update documentation accordingly
