+++ conf/config.yaml
@@ -16,7 +16,7 @@ model:
     gguf_dir: "${paths.gguf_directory}" # if you change this, change in run_pipeline
     use_8bit: false
   rkllm:
-    enabled: false # Enable RKLLM conversion
+    enabled: true # Enable RKLLM conversion
     target_platform: "rk3588" # Target Rockchip platform: rk3588, rk3576, etc.
     quantization: "w8a8" # RKLLM quantization: w8a8, w4a16, w4a16_g128
     output_dir: "rkllm_models" # Output directory for RKLLM models
@@ -38,8 +38,8 @@ training:
   logging_steps: 5
   warmup_steps: 10
   learning_rate: 2e-5
-  fp16: false
-  bf16: true
+  fp16: true
+  bf16: false
   weight_decay: 0.05
   max_seq_length: 2048
   optim: "paged_adamw_8bit"
@@ -54,6 +54,10 @@ training:
   optuna_n_trials: 10
   seed: 42
 
+data_preparation:
+  method: "classic" # Options: "classic" | "game"
+  # classic: Simple instruction/output format with basic prompting
+  # game: Complex game format with history, actions, and system messages
 
 grpo:
   val_data: "test_ru.json"
@@ -106,9 +110,12 @@ other:
   cutoff_len: 2048
   hf_login: false
 
+environment:
+  use_dotenv: false  # Set to true for local development, false for Docker/production
+
 testing:
   use_separate_files: false
-  test_split_ratio: 0.1
+  test_split_ratio: 0.05
   manual_lmstudio_test: false # use only for local run
   test: false
   test_dataset: "${paths.data_dir}/test_ru.json"
@@ -124,7 +131,7 @@ logging:
 
   mlflow:
     experiment_name: "vika-experiment"
-    tracking_uri: "http://127.0.0.1:5000"
+    tracking_uri: "http://mlflow:5000"
 
 
 hydra:
