+++ testing_model/__init__.py
@@ -3,10 +3,13 @@ from evaluation.deepeval_integration import (
     test_mention_number_of_values,
 )
 from evaluation.model_evaluation import test_llm, test_via_llamacpp
+from testing_model.models import CustomLocalModel, CustomMistralModel
 
 __all__ = [
     "test_llm",
     "test_mention_number_of_values",
     "test_from_dataset",
     "test_via_llamacpp",
+    "CustomLocalModel",
+    "CustomMistralModel",
 ]
